+++ packages/router/src/__tests__/links.test.tsx
@@ -2,9 +2,8 @@ import React from 'react'
 
 import { render } from '@testing-library/react'
 
-import { NavLink, useMatch, Link } from '../links'
+import { NavLink } from '../links'
 import { LocationProvider } from '../location'
-import { flattenSearchParams } from '../util'
 
 function createDummyLocation(pathname: string, search = '') {
   return {
@@ -277,79 +276,5 @@ describe('<NavLink />', () => {
     )
 
     expect(getByText(/Dunder Mifflin/)).not.toHaveClass('activeTest')
-  })
-})
-
-describe('useMatch', () => {
-  const MyLink = ({
-    to,
-    ...rest
-  }: React.ComponentPropsWithoutRef<typeof Link>) => {
-    const [pathname, queryString] = to.split('?')
-    const matchInfo = useMatch(pathname, {
-      searchParams: flattenSearchParams(queryString),
-    })
-
-    return (
-      <Link
-        to={to}
-        style={{ color: matchInfo.match ? 'green' : 'red' }}
-        {...rest}
-      />
-    )
-  }
-
-  it('returns a match on the same pathname', () => {
-    const mockLocation = createDummyLocation('/dunder-mifflin')
-
-    const { getByText } = render(
-      <LocationProvider location={mockLocation}>
-        <MyLink to="/dunder-mifflin">Dunder Mifflin</MyLink>
-      </LocationProvider>
-    )
-
-    expect(getByText(/Dunder Mifflin/)).toHaveStyle('color: green')
-  })
-
-  it('returns a match on the same pathname with search parameters', () => {
-    const mockLocation = createDummyLocation(
-      '/search-params',
-      '?page=1&tab=main'
-    )
-
-    const { getByText } = render(
-      <LocationProvider location={mockLocation}>
-        <MyLink to={`/search-params?tab=main&page=1`}>Dunder Mifflin</MyLink>
-      </LocationProvider>
-    )
-
-    expect(getByText(/Dunder Mifflin/)).toHaveStyle('color: green')
-  })
-
-  it('does NOT receive active class on different path', () => {
-    const mockLocation = createDummyLocation('/staples')
-
-    const { getByText } = render(
-      <LocationProvider location={mockLocation}>
-        <MyLink to="/dunder-mifflin">Dunder Mifflin</MyLink>
-      </LocationProvider>
-    )
-
-    expect(getByText(/Dunder Mifflin/)).toHaveStyle('color: red')
-  })
-
-  it('does NOT receive active class on the same pathname with different parameters', () => {
-    const mockLocation = createDummyLocation(
-      '/search-params',
-      '?tab=main&page=1'
-    )
-
-    const { getByText } = render(
-      <LocationProvider location={mockLocation}>
-        <MyLink to={`/search-params?page=2&tab=main`}>Dunder Mifflin</MyLink>
-      </LocationProvider>
-    )
-
-    expect(getByText(/Dunder Mifflin/)).toHaveStyle('color: red')
   })
 })
